Fix last known issue with Java 6.  We should be able to compile with either Java 5 or Java 6 compiler and run the resulting binary on either runtime.  <parallel> build of modules was somehow causing javac to choke on unboxing syntax (synchronization problem in the compiler?).  Building serially resolved the issue, but slowed the build significantly.  Better solution was to fork javac on all compiles.  Had to invoke javac directly on _jsp.java instead of relying on jasper2 (no way to specify fork="true" using jasper2 task).
Remove some explicit calls to intValue(), booleanValue(), etc. that I added while investigating this issue.

Fix StudyBvtTest to not rely on order of elements returned by getFormElementNameByTableCaption(), since this is unpredictable.

Don't let the crawler visit the memTracker page, where it's likely to find forbidden terms buried in stack traces.

Improve GO annotation loading error handling (null check before closing connection) and display correct filter text on GO charts.

Register a "study permissions" ViewFactory with permissions page instead of hard-coding study ui.  Correct bad wording and typo ("the the").

Start of improved caBIG admin ui.

3171 : Compiling under Java 6 fails due to unboxing issues
3105 : Typo on permissions page

+++ flow/src/org/labkey/flow/FlowPreference.java
@@ -44,7 +44,7 @@ public enum FlowPreference
     {
         try
         {
-            return Integer.valueOf(getValue(request)).intValue();
+            return Integer.valueOf(getValue(request));
         }
         catch (Exception e)
         {
